{"schema":"decodex/commit/1","summary":"Split live capture adapter modules","authority":"manual"}
